Automated rollback of changelist 181338957.
*** Reason for rollback ***

broke one of the tests.

*** Original change description ***

For low mem devices, running the dex2oat process in mini_boot can cause the
kernel to panic if there's not enough free memory for dex2oat process to write
the output file. In mini_boot mode, always assume the RAM to be 4GB if it's
lesser than that.

PiperOrigin-RevId: 182395042