Workaround non-threadsafe setenv code in android-qemu.
This envvar is set after android-qemu has started a bunch of threads and it
leads to segfaults when a call to getenv happens at the same time (seems
associated with mktime - which reads "TZ"  environment var which is a very common fn).

Leverage our preexisting ld_preload hack :( and unified_launcher to work around
this for now.
https://rachelbythebay.com/w/2017/01/30/env/

PiperOrigin-RevId: 189637129