Add an enable_test_harness flag
This adds a flag to control whether the ro.test_harness flag is set on a
device. This property completely disables package verification on the
device, which is problematic when you are trying to test package
verification.

Since this is a read-only property that can't be changed once set, it
is no longer set at build time as part of generating the ramdisk, but
rather in _StartEmulator, just after the device is started.

PiperOrigin-RevId: 190247179