Deflake openActionBarOverflowOrOptionsMenu by waiting for the next frame.
We need to wait for Activity#onPrepareOptionsMenu to be called before trying to open overflow or it's missing. onPrepareOptionsMenu is called by Choreographer after onResume and view is attached to window. To ensure the options menu is created before we try to open, wait for all processing tasks in the current frame to be finished.

PiperOrigin-RevId: 242723327