Use hyphen instead of underscore as device name separator.
When generating target names for tests which run on multiple devices, use
hyphen '-' instead of an underscore '_' to separate the device qualifier from
the rest of the target name. eg use
FooTest-emulated_generic_phone_29_x86

This helps distinguish the device name, and will also be useful when
attempting to query data by device type.

PiperOrigin-RevId: 307632964