Update AXT bazel to the protobuf 3.11.3 javalite toolchain
This is required to support Bazel 3.x. In Bazel 3.x, the implicit dependency to
@com_google_protobuf_javalite//:javalite_toolchain is gone, and you have to
switch to rules_proto, which depends on the javalite toolchain in protobuf
v3.11.3. Update all protolite runtime deps to
@com_google_protobuf//:protobuf_javalite, which is in 3.11.3. Also, update
the @com_google_protobuf_javalite alias to point to protobuf 3.11.3 in order
to maintain compatibility with Bazel 2.x and below.

PiperOrigin-RevId: 310412938