Minimal-case repro test for view scaling issue when using Espresso in Robolectric
Note that the issue only manifests when increasing the scale of the view. This is unsurprising because decreasing the scale shrinks the view, which means that any comparative size calculation would be trivial true.

What's especially interesting is that in this CL, the scale goes 1 -> 1.5 -> 0.5 -> 1. However, if you flip that order to go 1 -> 0.5 -> 1.5 -> 1, the test passes.

PiperOrigin-RevId: 339553167