Truncate large failure messages and stack traces.
Test failure events are sent over Binder, which have a 1MB transaction limit.
Large stack traces and messages can cause the parcel to exceed the 1MB limit,
resulting in the details being lost.

This change, similar to an earlier one made for AJUR and orchestrator v1, truncates
the message and stack to 65K.

PiperOrigin-RevId: 376182727