Workaround for backwards compatibility for monitor-1.4.0.
monitor-1.4.0-beta01 is not compatible with older releases of
androidx.test.core.

ActivityInvoker#getIntentForActivity is an interface method with a default
implementation, which is only supported in java8. The current androidx test
release build process will desugar this to java7 bytecode, which results
in a generated class like ActivityInvoker$$CC.

For some as yet undetermined reason, in the latest release the desugar
process will use ActivityInvoker$-CC as the class name. Thus older
androidx.test.core releases that call ActivityInvoker$$CC will get a
NoClassDefFoundError when using the latest monitor.

This change workarounds the problem by explicitly creating a class with
named ActivityInvoker$$CC.

PiperOrigin-RevId: 377533558