Fix IllegalAccessException in MODERN_ASYNC_TASK_EXTRACTOR
Due to Robolectric not instrumenting androidx.* classes by default any more
(which makes all classes public), the following error started appearing during
Robolectric tests that use Espresso:

    cannot access a member of class androidx.loader.content.ModernAsyncTask
    with modifiers "public static final"

Update `MODERN_ASYNC_TASK_EXTRACTOR` to use `getDeclaredField` and
`setAccessible(true)` when accessing the final `THREAD_POOL_EXECUTOR` field.

The problem is that ModernAsyncTask is not a public class, it's package
private. Trying to access members of a package private class outside of the
package is not allowed (JLS §6.6.1).

PiperOrigin-RevId: 384244512