Fix looper thread error in UiControllerImpl.
UiController was suspectible to 'java.lang.RuntimeException: Can't create
handler inside thread that has not called Looper.prepare()' errors if it
happened to be initialized off the main thread. Which can occur if
interruptEspressoTasks was called relatively early.

Eliminate the unnecessary separate initialize method entirely, and
make the Handler always use the passed in looper from the constructor.

PiperOrigin-RevId: 401385636