Remove unnecessary warning log
This `UiController` implementation is essentially a wrapper around Robolectric's controller, Robolectric runs the test on the main thread and doesn't need to be interrupted to stop main thread interaction, this should not be logged as a warning it should just be ignored. Previously every Espresso interaction in Robolectric resulted in this warning being subsequently printed.

PiperOrigin-RevId: 439875259