Fix the default PlatformTestStorage impl.
With Espresso 3.5.0, espresso will attempt to save a screenshot on test failures using PlatformTestStorage.
When the test services apk is not installed, PlatformTestStorage falls back to a simple FileTestStorage
impl that tries to write to raw filesystem, treating each outputpath as an absolute file path.

This commit borrows the androidx.benchmark heuristic for choosing an appropriate parent directory:
1. If additionalTestOutputDir is provided as a instrumentation argument
2. Else if API >=29 and a media store directory is available
3. Else choose app's externalCacheDir if set
4. else choose app's cacheDir

This commit also fixes the implementation for openInputFile to read from test apk assets,
and fixes getInputArg[s] to read from InstrumentationRegistry arguments.

PiperOrigin-RevId: 494751081