Indicates that a specific test or class should or should not be run on certain API levels.
Test(s) will be skipped when executed on android platforms less/more than specified level (inclusive) or if executed on android platforms that are included in the excludedSdks list.

If {@code @SdkSuppress} is applied at both the class and test method, the test method annotation takes precedence, and the class level {@code @SdkSuppress} is ignored.

PiperOrigin-RevId: 519236856