Wait up to 2 seconds for activity finisher to run.
A previous commit removed the waitForIdleSync after the activityFinisher was posted
in order to prevent deadlocks. However this could have side effects where it could
run after the next test has started, effectively inadvertently finishing an activity mid test.

This commit modifies the logic to wait up to 2 seconds for the activityFinisher to execute.

PiperOrigin-RevId: 525808008