Force initialization of instrumentationResultPrinter if its null.
Currently AJUR has hard coded per-SDK logic of when to initialize  instrumentationResultPrinter.

This commit simplifies the logic to just always attempt initialization
in either newApplication or onCreate. This should make AJUR more resilient
if the application lifecycle callback sequence ever changes, as it is speculated it does when instrumenting the system server itself.

PiperOrigin-RevId: 558162950