Update axt_versions.bzl to reflect the next published versions.
Previously, the versions listed in axt_versions.bzl represented the
last published versions of the artifact. This is problematic
since the gradle emulator tests were actually always testing
the last published versions, not the HEAD versions.

This commit updates axt_versions.bzl to reflect the next
planned version. Logically this makes more sense as the version
string can be updated from micro to minor when there are API changes.