Replace python logic to build combined maven zip with kotlin.
Previously this repo used a simple python script to produce a combined zip of maven
artifacts. However this is the only use of python in the project, and it required
bringing in a python bazel rules dependency.

Requiring an extra dependency just to combine zips was overkill,
especially since there is existing kotlin logic to combine jars which is
very similar.

Also refactor JarCombiner to separate the logic for duplicate detection and
stripping detection.

PiperOrigin-RevId: 563178825