Use gradle constants consistently and use api 33 for target device.
This will upgrade the runner, orchestrator and espresso projects to compile/target sdk 34,
and upgrade all projects to use the aosp-atd 33 image.

Previously espresso.accessibility needed to use the non-atd image because atd 30 has an old bug
where browser intents are not supported. This should be fixed on 33 thus allowing all tests
to use the same image, which should speed up runtime.

PiperOrigin-RevId: 581011723