Updates captureToBitmap in ViewCapture to treat API 34 the same way that all previous API works.
Because API 34 wasn't getting the surface, it was including the action bar and messing up the screenshots. By removing it, it will use the surface instead and it works perfectly.

API 28 still does not work right with dialog boxes, but that is a known issue that we don't have a fix for yet.

To test this, I added tests for every API level and fixed a bug where the action bar wasn't showing up in the test.

PiperOrigin-RevId: 588915479