Stop posting unnecessary work to threads in espresso.
Currently Espresso will post no-op empty runnables to a background thread pool for every interaction
when running Espresso in a normal, non-remote environment.

This is wasteful and causes some noise for leaked-thread detector facilities.

This commit addresses this by adding an extra API to RemoteInteraction, so it only posts the remote
commumnication runnable if its running in remote mode.

PiperOrigin-RevId: 600526081