Treat exceptions during DefaultFailureHandler screenshots as suppressed exceptions
If an Espresso failure occurs, DefaultFailureHandler will capture a window
snapshot and take a screenshot.

However, if a RuntimeException occurs during this process, it will replace the
original exception.

An example is if there is an exception that occurs during a View's rendering
logic. The original exception will trigger DefaultFailureHandle's screenshot,
which may trigger a different exception, hiding the root cause.

PiperOrigin-RevId: 601202589