Cache Gradle Managed Devices emulator snapshot.
It currently takes ~50 seconds on each CI run to boot the emulator
and save a snapshot.

This commit saves the snapshot to a github action cache, to save the
cost of booting on each run.

However, compressing and saving the snapshot to cache is also expensive (~30 seconds)
so we don't want to do this for every run. Ideally there would be a precise
mechanism for determining if the snapshot has changed. For now, just use a cache
key encorporating the date and the hash of the top level build.gradle file, so
the cache is generated every week.