Split by whitespace for commands
This is needed because the ShellExecutor uses ProcessBuilder which is known to have encoding issues as
it converts Java Strings to C-strings (byte arrays) with no regards to encoding. In order to avoid, the
best we can do here is to avoid using characters which can cause encoding issues (eg:whitespaces).

Empty-strings also causes the same behavior for which we will raise an IllegalStateException to protect
from that as debugging the unusual behavior is time-consuming without any idea what's going wrong so
crashing early will make it much easier.

PiperOrigin-RevId: 611160643