Reduce TestStorage logspam.
https://github.com/android/android-test/commit/5acf36a9ae2f59d20af8b7adf4c9b27601f58dbe added
logging whenever TestStorage calculates a file path, which can occur several times in a test run.
This is unnecessary and spams logcat.

This commit removes most logging, except for the special case where a new directory is chosen
when running as a non system user. And in that case a check is added to ensure the log
only happens once.

PiperOrigin-RevId: 630169476