Refactor Interrogator MessageQueue access to a TestLooperManagerCompat class.
An upcoming TestLooperManager change will add the ability for Espresso's Interrogator class
to use it directly, replacing the use of reflection to access MessageQueue internals.

This commit prepares for this change by refactoring MessageQueue access to a
TestLooperManagerCompat container class. A future change will enhance
TestLooperManagerCompat to use the public TestLooperManager API when the Android
runtime environment supports it.

PiperOrigin-RevId: 706812503