Fix Espresso.onIdle for non-main Loopers on Baklava.
LooperIdlingResourceInterrogationHandler, which is used for non main Loopers,
requires calling peekAtQueueState from the main Looper thread. The previous Interrogator implementation,
when used on Baklava, didn't effectively support this because TestLooperManager.acquire
needs to be called from the Looper's thread.

This commit refactors Interrogator so it be acquired and stored on the Looper's current thread.

PiperOrigin-RevId: 714149090