Avoid potentially exception throwing side effects from `IdlingResourceTimeoutException` being thrown
Interactions with `TestOutputEmitter` can throw, and if it does it suppresses the `IdlingResourceTimeoutException` exception from even being created. Instead of having this side effect in the `IdlingResourceTimeoutException` constructor move the thread dumping out of the constructor and wrap it in a try/catch, attaching any exceptions as suppressed by the `IdlingResourceTimeoutException` exception.

In particular this can happen if more than one `TestOutputHandler` is registered.

PiperOrigin-RevId: 802381773