Catch `Error` when dumping thread states in `IdlingPolicy`.
Expand the catch block to include `Error` types in addition to `RuntimeException` when calling `TestOutputEmitter.dumpThreadStates`. This ensures that any issues during thread state dumping are suppressed and added to the `AppNotIdleException`, preserving the original exception as the primary cause.

`java.lang.ExceptionInInitializerError` can happen when static initializing `TestOutputEmitter` if more than one service is discovered.

PiperOrigin-RevId: 802644040