BEGIN_PUBLIC Define isEqualTo for `PersistableBundleSubject`
Defines a brute force `isEqualTo` API for `PersistableBundle` by converting it to a map and taking advantage of the existing `MapSubject`. `PersistableBundle` has historically been very painful to test since they are lazy containers and do not implement equals. While the existing per value subjects are useful, they do not allow for generic comparison in parametrized tests which limits flexibility and results in more boilerplate in unit tests.
END_PUBLIC

PiperOrigin-RevId: 816295120