Update PersistableBundleSubject#isEqual to handle arrays
I originally thought MapSubject was handling `Array` types, but the PersistableBundles in the test cases were using the same array reference meaning the tests were incorrect.

Updates to just convert the arrays to a list and updates the tests to perform a clone in `getTestBundle`. We could also convert them to a function instead but a shallow copy is sufficient here.

PiperOrigin-RevId: 827559640